requirements.txt
pytest
playwright
python-dotenv
allure-pytest
